// src/shared/api/generated/types/CreatedDocumentShareLinkDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type CreatedDocumentShareLinkDto = {
  /**
   * @type number
   */
  id: number;
  /**
   * @type string
   */
  tokenPrefix: string;
  /**
   * @type string, date-time
   */
  expiresAt?: string | null;
  /**
   * @type string, date-time
   */
  revokedAt?: string | null;
  /**
   * @type string, date-time
   */
  createdAt: string;
  /**
   * @type string, date-time
   */
  lastAccessAt?: string | null;
  /**
   * @type number
   */
  accessCount: number;
  /**
   * @type boolean
   */
  isActive: boolean;
  /**
   * @type boolean
   */
  isExpired: boolean;
  /**
   * @type number
   */
  versionId: number;
  /**
   * @type number
   */
  versionNumber: number;
  /**
   * @type number
   */
  documentId: number;
  /**
   * @type string
   */
  documentTitle: string;
  /**
   * @description Returned only by the create operation
   * @type string
   */
  token: string;
  /**
   * @type string
   */
  sharePath: string;
};