// src/shared/api/generated/types/DocumentShareLinkDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type DocumentShareLinkDto = {
  /**
   * @type number
   */
  id: number;
  /**
   * @type string
   */
  tokenPrefix: string;
  /**
   * @type string, date-time
   */
  expiresAt?: string | null;
  /**
   * @type string, date-time
   */
  revokedAt?: string | null;
  /**
   * @type string, date-time
   */
  createdAt: string;
  /**
   * @type string, date-time
   */
  lastAccessAt?: string | null;
  /**
   * @type number
   */
  accessCount: number;
  /**
   * @type boolean
   */
  isActive: boolean;
  /**
   * @type boolean
   */
  isExpired: boolean;
  /**
   * @type number
   */
  versionId: number;
  /**
   * @type number
   */
  versionNumber: number;
  /**
   * @type number
   */
  documentId: number;
  /**
   * @type string
   */
  documentTitle: string;
};